bumps up the timeout to 60s
This was a result of a discussion on diego's slack channel. some users
reported an issue with small instances taking too long to setup the certs
causing the script to timeout prematurely. It is not clear whether this
is the right course of action, but this is a stopgap solution.

Signed-off-by: John Shahid <jvshahid@gmail.com>